Integration test for alignment constraints
* Don't try to convert key partitions when the transformation is not convertible

* Minor fix for the return type for MachineDesc

* Fix broken integration tests

* Integration test for alignment constraints

* Increase parallelism for test build

* Start testing with multiple CPUs and in test mode

See merge request legate/legate.core.internal!40

+++ src/core/mapping/machine.cc
@@ -187,9 +187,9 @@ std::vector<TaskTarget> MachineDesc::valid_targets_except(
   return result;
 }
 
-size_t MachineDesc::count() const { return count(preferred_target); }
+uint32_t MachineDesc::count() const { return count(preferred_target); }
 
-size_t MachineDesc::count(TaskTarget target) const { return processor_range(target).count(); }
+uint32_t MachineDesc::count(TaskTarget target) const { return processor_range(target).count(); }
 
 std::string MachineDesc::to_string() const
 {
